Version 2: 02/23/23 GG
Corrected target heading calculation to be normalized to a compass bearing. Currently commented out SetVelocity() function calls as it is not working. Hardcoded the driving speed for TurnToHeading() function. GPS navigation seems to drive past GPS waypoint without stopping. We would expect it to drive circles around the waypoint as it is written.