Version 4: 03/02/2023 GG
Made several QOL changes to format such as removing global variables that didn't need to be global, updating comments to be more accurate and descriptive, and removing the SetVelocity() function as it was not being used.

Also changed DeadManSwitch() so that it was clearer in the principle file how it was toggling between modes.

Moved GPSNavigation() to the principle file also for clarity, and updated it so that it could loop through a series of target coordinates.