Version 5.1: 03/10/23 G
Added preprocessor directives to turn on debugging prints for the serial monitor and bluetooth telemetry. Adjusted the behavior of the collision detection functions in an attempt to clean up future references to collision detection in other functions. Also added Leo's adjustment to the parsing function that correctly applies negative  values to South and West directions.